refactor: factor private routes into own file
